feat - [async handler and error middleware in backend]

+++ backend/controllers/deleteVector.controller.ts
@@ -1,16 +1,24 @@
-import { Request, Response } from 'express';
-
 import { pineconeClientInit } from '../lib/pinecone';
+import asyncHandler from '../handlers/async.handler';
+import ErrorHandler from '../handlers/error.handler';
 
-const deleteAllVectors = async (req: Request, res: Response) => {
-    const { nameSpace } = req.body;
-    const pinecone = await pineconeClientInit();
-    const index = pinecone.index(process.env.PINECONE_INDEX_NAME);
-    const ns = index.namespace(nameSpace);
-    await ns.deleteAll();
-    res.json({
-        success: true,
-    });
-};
+import type { Request, Response, NextFunction } from 'express';
+
+const deleteAllVectors = asyncHandler(
+    async (req: Request, res: Response, next: NextFunction) => {
+        const { nameSpace } = req.body;
+        if (!nameSpace) {
+            return next(new ErrorHandler(400, 'NameSpace is required'));
+        }
+        const pinecone = await pineconeClientInit();
+        const index = pinecone.index(process.env.PINECONE_INDEX_NAME);
+        const ns = index.namespace(nameSpace);
+        await ns.deleteAll();
+        res.json({
+            success: true,
+            message: 'All vectors deleted successfully',
+        });
+    },
+);
 
 export { deleteAllVectors };
